refactor: fold stored payload module into record
