+++ storage/src/record/encryption.rs
@@ -35,10 +35,17 @@ use rand::random;
 use s2_common::{
     deep_size::DeepSize,
     encryption::{EncryptionAlgorithm, EncryptionSpec},
-    record::{EnvelopeRecord, Metered, MeteredExt as _, MeteredSize, Record, SeqNum},
+    record::{EnvelopeRecord, Metered, MeteredExt as _, MeteredSize, Record, SeqNum, Sequenced},
+    stream::{
+        AppendInput, AppendRecord, AppendRecordBatch, AppendRecordParts, ReadBatch,
+        ReadSessionOutput,
+    },
 };
 
-use super::{Encodable, StoredRecord, StoredRecordDecodeError, decode_envelope_record};
+use super::{
+    Encodable, StoredAppendInput, StoredReadBatch, StoredReadSessionOutput, StoredRecord,
+    StoredRecordDecodeError, decode_envelope_record,
+};
 
 const FORMAT_ID_LEN: usize = 1;
 
@@ -246,6 +253,36 @@ pub fn encrypt_record(
     record.metered()
 }
 
+pub fn encrypt_append_input(
+    input: AppendInput,
+    encryption: &EncryptionSpec,
+    aad: &[u8],
+) -> StoredAppendInput {
+    let AppendInput {
+        records,
+        match_seq_num,
+        fencing_token,
+    } = input;
+    let records = records
+        .into_iter()
+        .map(|record| {
+            let AppendRecordParts { timestamp, record } = record.into_parts();
+            AppendRecord::try_from(AppendRecordParts {
+                timestamp,
+                record: encrypt_record(record, encryption, aad),
+            })
+            .expect("record encryption preserves append record limits")
+        })
+        .collect::<Vec<_>>();
+
+    AppendInput {
+        records: AppendRecordBatch::try_from(records)
+            .expect("record encryption preserves append batch limits"),
+        match_seq_num,
+        fencing_token,
+    }
+}
+
 fn prep_encryption_buffer(
     envelope: &EnvelopeRecord,
     format: EncryptedRecordFormat,
@@ -314,6 +351,40 @@ pub fn decrypt_stored_record(
             Ok(record.metered())
         }
     }
+}
+
+pub fn decrypt_read_session_output(
+    output: StoredReadSessionOutput,
+    encryption: &EncryptionSpec,
+    aad: &[u8],
+) -> Result<ReadSessionOutput, RecordDecryptionError> {
+    match output {
+        ReadSessionOutput::Heartbeat(tail) => Ok(ReadSessionOutput::Heartbeat(tail)),
+        ReadSessionOutput::Batch(batch) => {
+            decrypt_read_batch(batch, encryption, aad).map(ReadSessionOutput::Batch)
+        }
+    }
+}
+
+fn decrypt_read_batch(
+    batch: StoredReadBatch,
+    encryption: &EncryptionSpec,
+    aad: &[u8],
+) -> Result<ReadBatch, RecordDecryptionError> {
+    let records: Result<Metered<Vec<Sequenced<Record>>>, RecordDecryptionError> = batch
+        .records
+        .into_inner()
+        .into_iter()
+        .map(|record| {
+            let (position, record) = record.into_parts();
+            decrypt_stored_record(record, encryption, aad).map(|record| record.sequenced(position))
+        })
+        .collect();
+
+    Ok(ReadBatch {
+        records: records?,
+        tail: batch.tail,
+    })
 }
 
 fn decrypt_payload(
